Fix 5 bugs: unwrap in prod, silent graph errors, fusion weights, Gini NaN, f32/f64 inconsistency
- source.rs: Replace unwrap() with if-let in batch token flushing (CLAUDE.md violation)
- node.rs: Surface 9 silent `let _ =` graph sidecar errors as tracing::warn (merge/correct/split)
- fusion.rs: Log warning on weight/list length mismatch instead of silent fallback
- skewroute.rs: Filter NaN/Inf values before Gini coefficient computation
- traits.rs + postgres/source.rs: Change SourceRepo::update_embedding from &[f32] to &[f64] for
  consistency with all other repos (NodeRepo, ChunkRepo, ArticleRepo)
- 4 new tests for fusion weight mismatch and Gini NaN handling

+++ engine/crates/covalence-core/src/storage/postgres/source.rs
@@ -174,7 +174,7 @@ impl SourceRepo for PgRepo {
         Ok(row.get("count"))
     }
 
-    async fn update_embedding(&self, id: SourceId, embedding: &[f32]) -> Result<()> {
+    async fn update_embedding(&self, id: SourceId, embedding: &[f64]) -> Result<()> {
         let pgvec = format!(
             "[{}]",
             embedding
